corpus: refresh the banked test artifacts against the rebanked tree
The 43 banked packages' test sources, test-info anchors, hosts and
tests.csproj files as regenerated by the campaign's own validation
sweep - each provably paired with the production tree it validated
against (43/43). Includes the newly-written external-variant
package_info_external_test.cs files (math, math/bits) from the
TypeAccessibility external-variant rule.

+++ src/go-src-converted/container/heap/heap_test.cs
@@ -67,7 +67,7 @@ public static void TestInit0(ж<testing.T> Ꮡt) {
         nint x = Pop(new myHeapжInterface(h))._<nint>();
         (~h).verify(Ꮡt, 0);
         if (x != 0) {
-            Ꮡt.Errorf("%d.th pop got %d; want %d"u8, i, x, 0);
+            Ꮡt.Errorf("%d.th pop got %d; want %d"u8, i, x, (nint)(0));
         }
     }
 }
